+++ bika/lims/catalog/analysisrequest_catalog.py
@@ -20,6 +20,7 @@ _indexes_dict = {
     # TODO: Can be removed? Same as id
     'sortable_title': 'FieldIndex',
     'getClientUID': 'FieldIndex',
+    'getClientID': 'FieldIndex',
     'cancellation_state': 'FieldIndex',
     'getBatchUID': 'FieldIndex',
     'getDateSampled': 'DateIndex',
@@ -62,6 +63,7 @@ _columns_list = [
     'getBatchURL',
     'getClientUID',
     'getClientTitle',
+    'getClientID',
     'getClientURL',
     'getContactUID',
     'getContactUsername',